+++ src/views/Favorites.vue
@@ -28,20 +28,27 @@
                             : 'primary--text text--darken-2'
                     "
                 >
-                    <v-tab>
-                        {{ $t("views.home.liveOrUpcomingHeading") }}
-                        <v-chip small class="ml-1 px-2" color="primary">{{ lives.length + upcoming.length }}</v-chip>
+                    <v-tab class="pa-2">
+                        {{ $t("views.home.liveOrUpcomingHeading").split(/[\/・]/)[0].trim() }}
+                        <span class="stream-count-chip mx-1 rounded-md primary white--text rounded-lg pa-1">
+                            {{ lives.length }}
+                        </span>
+                        /
+                        {{ $t("views.home.liveOrUpcomingHeading").split(/[\/・]/)[1].trim() }}
+                        <span class="stream-count-chip ml-1 rounded-md primary white--text rounded-lg pa-1">
+                            {{ upcoming.length }}
+                        </span>
                     </v-tab>
-                    <v-tab>
+                    <v-tab class="pa-2">
                         {{ $t("views.home.recentVideoToggles.official") }}
                     </v-tab>
-                    <v-tab>
+                    <v-tab class="pa-2">
                         {{ $t("views.home.recentVideoToggles.subber") }}
                     </v-tab>
                 </v-tabs>
             </portal>
             <LoadingOverlay :isLoading="false" :showError="hasError" />
-            <template v-show="!hasError">
+            <div v-show="!hasError">
                 <template v-if="tab === Tabs.LIVE_UPCOMING">
                     <SkeletonCardList v-if="isLoading" :cols="colSizes" :dense="currentGridSize > 0" />
                     <template v-else-if="lives.length || upcoming.length">
@@ -92,7 +99,7 @@
                         </generic-list-loader>
                     </keep-alive>
                 </template>
-            </template>
+            </div>
         </template>
         <template v-else>
             <div class="ma-auto d-flex flex-column align-center">