+++ src/views/Home.vue
@@ -25,20 +25,27 @@
                     $store.state.settings.darkMode ? 'primary--text text--lighten-3' : 'primary--text text--darken-2'
                 "
             >
-                <v-tab>
-                    {{ $t("views.home.liveOrUpcomingHeading") }}
-                    <v-chip small class="ml-1 px-2" color="primary">{{ lives.length + upcoming.length }}</v-chip>
+                <v-tab class="pa-2">
+                    {{ $t("views.home.liveOrUpcomingHeading").split(/[\/・]/)[0].trim() }}
+                    <span class="stream-count-chip mx-1 rounded-md primary white--text rounded-lg pa-1">
+                        {{ lives.length }}
+                    </span>
+                    /
+                    {{ $t("views.home.liveOrUpcomingHeading").split(/[\/・]/)[1].trim() }}
+                    <span class="stream-count-chip ml-1 rounded-md primary white--text rounded-lg pa-1">
+                        {{ upcoming.length }}
+                    </span>
                 </v-tab>
-                <v-tab>
+                <v-tab class="pa-2">
                     {{ $t("views.home.recentVideoToggles.official") }}
                 </v-tab>
-                <v-tab>
+                <v-tab class="pa-2">
                     {{ $t("views.home.recentVideoToggles.subber") }}
                 </v-tab>
             </v-tabs>
         </portal>
         <LoadingOverlay :isLoading="false" :showError="hasError" />
-        <template v-show="!hasError">
+        <div v-show="!hasError">
             <template v-if="tab === Tabs.LIVE_UPCOMING">
                 <SkeletonCardList v-if="isLoading" :cols="colSizes" :dense="currentGridSize > 0" />
                 <template v-else-if="lives.length || upcoming.length">
@@ -89,7 +96,7 @@
                     </generic-list-loader>
                 </keep-alive>
             </template>
-        </template>
+        </div>
     </v-container>
 </template>
 
@@ -247,5 +254,10 @@ export default {
 <style>
 .v-slide-group__prev--disabled {
     display: none !important;
+}
+/* shared with favorites.vue */
+.stream-count-chip {
+    letter-spacing: normal;
+    min-width: 24px;
 }
 </style>