+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Holodex Changelog
 
+### 2.11.8 Holodex [*June 9, 2021*]
+
+##### Patches
+- In honor of Coco-chan, we added a **Coco theme** to Holodex!
+- [You can now drag and drop videos to MultiView cells from youtube and holodex](https://github.com/RiceCakess/Holodex/pull/281)
+- [Scrollbar and Gutter adjustment for MultiView](https://github.com/RiceCakess/Holodex/pull/280)
+- Fixed bug when you use custom presets and you delete a 2-video preset and it doesn't go to a 1-video preset.
+- Fixed the system shunting you over to Favorites main page by default after a while. Oops.
+
+##### Org changes
+- (+) WACTOR org has been added
+
 ### 2.11.7 Holodex [*June 1, 2021*]
 
 > Devs are on break! (Well one of them is just too busy playing factorio) 😁 However, we're still accepting pull requests from the community!